third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGLinearGradientElement-dom-gradientUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'gradientUnits' attribute of the SVGLinearGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linearGradientElement.getAttribute('gradientUnits') is null
PASS linearGradientElement.getAttribute('gradientUnits') is "userSpaceOnUse"
PASS successfullyParsed is true

TEST COMPLETE
